Updated the OC function. The function is 100% working at this point. Only need to get the check function working

+++ cpp_code/check.h
@@ -2,10 +2,11 @@
 #define CHECK_h
 
 #include <stddef.h>
+#include <stdlib.h>
 #include "Eigen"
 
 using namespace Eigen;
 
-MatrixXd check(int nelx, int nely, int rmin, MatrixXd x, MatrixXd dc);
+MatrixXd check(int nelx, int nely, double rmin, MatrixXd x, MatrixXd dc);
 
 #endif
